Move TaskList component up in directory tree.
