+++ client/dashboard/customizable.js
@@ -23,7 +23,7 @@ import defaultSections from './default-sections';
 import Section from './section';
 import withSelect from 'wc-api/with-select';
 import { recordEvent } from 'lib/tracks';
-import TaskList from './task-list';
+import TaskList from '../task-list';
 import { isOnboardingEnabled } from 'dashboard/utils';
 import {
 	getCurrentDates,
