+++ client/task-list/index.js
@@ -20,7 +20,7 @@ import { PLUGINS_STORE_NAME } from '@woocommerce/data';
  * Internal dependencies
  */
 import './style.scss';
-import CartModal from '../components/cart-modal';
+import CartModal from 'dashboard/components/cart-modal';
 import { getAllTasks } from './tasks';
 import { recordEvent } from 'lib/tracks';
 import withSelect from 'wc-api/with-select';
