+++ client/task-list/tasks/payments/images/wcpay.js
